babel and cli
For the cli programs to work they need to compiled babel. There may be
better ways of doing this but I seaprated out the react building from
the cli building using specific babel environments (cli and browser) and
then used the @babel/plugin-transform-runtime in the cli environment. It
seems to work after trying an install from npmjs but I'd like to
simplify it if possible.